pyproject.toml
[tool.poetry]
name = "espetofc-api"
version = "0.1.0"
description = ""
authors = ["Luis Henrique Carreira <luishenriquecarreira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
SQLAlchemy = "^2.0.30"
pydantic-settings = "^2.2.1"
psycopg2-binary = "^2.9.9"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
